refactor(core): extract path-scope detection from _hooklib (#321, slice 3/4)
Third of four. _hooklib goes 880 -> 663 lines. It was 1,263 before the partition
began, so three slices have taken nearly half off the god module.

_scopelib owns which repo paths are database migrations (H-14), CI/CD workflow
files (H-15), and deployment / IaC manifests (H-16), plus the
security-controls.md reader and glob compiler they share.

THE CLEANEST SEAM OF THE FOUR, and measured the same way as the others: this
cluster referenced NOTHING from the rest of _hooklib, and NOTHING in the rest of
_hooklib referenced it. Its only dependency was norm_path, already on the
_pathnorm floor. Zero edges in either direction.

They belong together because they are ONE MECHANISM wearing three labels: all
three predicates are path_in_globs(rel, root, <DEFAULTS>, <DECL_RE>) over the
same glob compiler, the same controls reader, and the same two mtime-keyed
caches. A change to how a project DECLARES a scope lands in one place instead of
three. The defaults differ; the machinery does not.

THE PRIVATE NAMES ARE RE-EXPORTED TOO, and that was not optional. The first cut
moved 13 underscore-prefixed symbols out and test_hooklib immediately failed on
`_hooklib._read_controls` and `_hooklib._MIGRATION_DEFAULT_RES`: seven of the
thirteen have real consumers. Same lesson slice 1 taught with reset_host - an
underscore prefix means "not public", never "unused". Checked the cache
semantics before sharing the binding: _CONTROLS_CACHE is mutated in place and
never rebound, so the re-export shares one cache rather than forking it.

Public surface verified against origin/main: nothing lost, nothing gained.
Parity rests on 1,185 pre-existing hook tests that did not move, plus the
cold-install matrix (285 assertions, ca + ca-codex).

ca 2.10.6 -> 2.10.7, ca-codex 0.3.5 -> 0.3.6, ca-pi 0.1.41 -> 0.1.42.
